+++ clodoo/check_4_seq.sh
@@ -1,4 +1,4 @@
-# __version__=0.3.31.4
+# __version__=0.3.31.5
 THIS=$(basename $0)
 if [ -z "$1" ]; then
   echo "$0 db [user]"