+++ clodoo/tests/clodoo_test_03.py
@@ -22,7 +22,7 @@ except BaseException:
     from clodoolib import build_odoo_param
 
 
-__version__ = "0.3.31.4"
+__version__ = "0.3.31.5"
 
 
 MODULE_ID = 'clodoo'