+++ clodoo/upd_oemod.py
@@ -12,7 +12,7 @@ except ImportError:
     import z0lib
 
 
-__version__ = "0.3.31.4"
+__version__ = "0.3.31.5"
 
 
 def initialize_params(ctx):