+++ wok_code/egg-info/history.rst
@@ -1,4 +1,9 @@
+0.1.17.3 (2021-07-15)
+~~~~~~~~~~~~~~~~~~~~
+
+* [FIX] wget_odoo_repositories.py: best debug mode: check for branch
+
 0.1.17.1 (2021-02-19)
 ~~~~~~~~~~~~~~~~~~~~~
 
-[IMP] wget_odoo_repositories.py: best debug mode
+* [IMP] wget_odoo_repositories.py: best debug mode